+++ setup.py
@@ -44,7 +44,7 @@ setup_args = dict(
     zip_safe=False,
     entry_points = {
         'pytest11': [
-            'pytest-importnb = importnb.test',
+            'pytest-importnb = importnb.pytest_plugin',
         ]
     },
 )
